Move PsuChassisInfo class tests to a new file, test_PsuChassisInfo.py
